requirements.txt
flask>=3.0.0
yt-dlp>=2024.1.0
requests>=2.31.0
pillow>=10.0.0
gunicorn>=21.2.0
flask-cors>=4.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
curl-cffi>=0.6.0
imagehash>=4.3.1
opencv-python-headless>=4.8.0
opencv-contrib-python-headless>=4.8.0
transformers>=4.38.0
torch>=2.2.0
